Remove inline (block) version of require_admin
This is only used in two places, and the places where it is used don't
really need any admin-only protection because we are already only
showing a publicly-visible subset of the the record data.

(In the case of issues, our before filters already will only find issues
for which the viewer has access rights, and in the case of links, all
links are public. In both cases, we return only a partial JSON hash
anyway.)

If we ever find a place where we actually need an inline/block version
of #require_admin, it is easily added back, but in the meantime it's
good to keep our controller actions as simple as possible. If we don't
need to clutter them up with a special block case, then we shouldn't.

Signed-off-by: Wincent Colaiuta <win@wincent.com>