Add emails#show template
This is really just an idea that I was trying out (abusing a form
without a submit button) so as to get free checkboxes and table-free
layout. In the end, I don't think I'll go with it, but I wanted to
commit it so it would at least appear in the repo history.

Signed-off-by: Wincent Colaiuta <win@wincent.com>