Unify confirm_email_and_redirect handling
In the Users controller we are now only doing the
"confirm_email_and_redirect" thing from one action, so the auxilliary
helper method is no longer required. (Note there was a bug in it anyway,
as the base_msg parameter had no effect).

The other place in the Users controller which used to need to do this
was the #update action, and that part of its responsibilities are now
handled by the Emails controller, so teach that how to send confirmation
messages as well.

Although almost the same code is now repeated in two places, it's not
complex enough to warrant keeping it DRY (either in the application
controller or in a module) as that would actually just make it harder to
maintain and understand.

Signed-off-by: Wincent Colaiuta <win@wincent.com>