Add CommentObserver
Much like we did with the Issue model, extract new comment alert sending
out into an observer class. This kind of detail is not really a core
concern of the model, so it is cleaner if it can be transferred over to
an observer.

There are a additional callbacks still in the Comment model that we may
still consider moving out into the observer also.

Signed-off-by: Wincent Colaiuta <win@wincent.com>