Hide #comment-form rather than replacing it
This is preliminary work to adding more AJAX to the commenting flow.
Instead of redirecting after a comment is added, we want to insert it
inline.

That in turn means that we need a way of putting the comment form back
into an initial state.

Note there are a number of style fixes in this commit in the surrounding
area, but only one functional change.

Signed-off-by: Wincent Colaiuta <win@wincent.com>