Extract comment rendering out into partial
This is a good thing to do, as it DRYs up 4 templates considerably.
There are still two more to be addressed (issue and topic comments)
because they use a different format (table-style instead of
speech-bubble style).

This will be useful shortly when further we AJAXify the commenting
system, because it will make it trivial to render individual or grouped
comments in response to new comment creations. (The two tricky issues
are getting this to work for both styles of comment, and possibly also
making sure that the #cycle calls in the partial make sense.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>