Fix whitespace bug in field_error_proc
This bug has been around ever since the field_error_proc was added, but
I just noticed it now while fixing the Rails 3.1.0.rc3 breakage. We were
inserting the space in the wrong place before the "field_with_errors"
class:

  <input class=" field_with_errorsother" ...>

Now we correctly insert it on the right instead of the left:

  <input class="field_with_errors other" ...>

Note that I chose to interpolate within a single string:

  %Q[\\1#{error_class} ]

Rather than concatenate strings at runtime:

  '\1' + error_class + ' '

And changed the second subsitution line to the same style for
consistency, even though it was bug-free.

The alternative would be to use an (almost) literal string for both
cases, but that would involve some repetion:

  '\1field_with_errors '
  '\1 class="field_with_errors"'

Signed-off-by: Wincent Colaiuta <win@wincent.com>