Make issues#show use comments partial
We're also enhancing the partial a little bit here, extracting the logic
used to deterimine the class into a helper. Note that we add the
"private" class here when appropriate, despite the fact that the
comments partial may end up being rendered in page-cached pages (note
that pages which have special content when viewed by an admin are not
page-cached, so we need not worry about stuff getting inappropriately
cached).

Signed-off-by: Wincent Colaiuta <win@wincent.com>