Generalize comment_class for use by topics
Topics are shown like comments, so tweak the comment_class helper so
that we can use it to generate the CSS class for topics as well. Note
that at the same time we DRY up the templates a little by getting the
comment class helper to prepend "comment" to the returned string.

Signed-off-by: Wincent Colaiuta <win@wincent.com>