Fix broken comments#update routing (partially reverts 4449c73)
Perhaps 4449c73 worked at the time as its commit log message indicates,
but right now it doesn't work (breaks form_for). So, revert it. Tested
both the edit form and the ham button.

The good news is that this makes both work. The bad news is that there
are some failing specs:

- if we support :update at the top, non-nested level, then our "ham"
  button works, but our specs fail
- this is because the router expects our generated routes to match
  against the top level, and our nested routes don't do that

So, need to think about what the nicest way to get out of that is. If I
teach the button_to_moderate_comment_as_ham method to generate nested
target URLs, I'll have an N+1 SELECT problem in the comments#index view
(because each comment will have to look up its commentable).

Signed-off-by: Wincent Colaiuta <win@wincent.com>