Don't specify BUNDLE_PATH in local .bundle/config
We're using RVM locally now because of segfaults in the version of Ruby
that comes with Mac OS X Lion. This in turn means that we don't need (or
really want) a specific bundle path set, as RVM will handle an
appropriate "sandbox" for us.

On the production site this change will have no effect as we already
only call "bundle install" with an explicit path.

Signed-off-by: Wincent Colaiuta <win@wincent.com>