Concatenate JS files
Using the assets pipeline without actually concatenating all the files
via a manifest is like installing new electrical wiring into a building
but continuing to use candles for illumination (or something like that).

So now we concatenate all JS files into one. Dynamic JS snippets
continue to be added on a per-controller basis, and also via the custom
JS controller.

Signed-off-by: Wincent Colaiuta <win@wincent.com>