Fix improper center of forms
All forms were displaying with text-align center, due to the collapsing
of all site styles into a single stylesheet.

Unfortunately, I cannot confirm that this actually works as the asset
pipeline is currently not cooperating (it's continuing to render the old
styles even when I blow away "public/assets/*" and "tmp/sass-cache").

Signed-off-by: Wincent Colaiuta <win@wincent.com>