Add index to needles table
This should make search performance a little better. Kind of amazing
that it worked viably at all, as it only had the primary key (index on
the id column). By indexing on (content, attribute_name) the index will
be used both by queries which limit by content (eg. "WHERE content =
'foo'") and those which limit by content and attribute_name (eg. "WHERE
content = 'foo' AND attribute_name = 'title'").

Note that we have no index on other things like `public` or `user_id`,
but I am expecting the index to be used to reduce the number of rows
that must be looked at anyway.

Signed-off-by: Wincent Colaiuta <win@wincent.com>