Replace akephalos with capybara-webkit
Akephalos was sweet while it lasted -- it still works! -- but the fact
that the official repo is abandonware is a real shame; I can only infer
that Bernard must have started using something else.

That something might be zombie.js, or it might be capybara-webkit. I
figured I'd try out the latter. So far, it looks great; the request
specs all still pass, but much faster. Here is the best of three runs
for `time rspec spec/requests` using Akephalos:

  Finished in 36.6 seconds
  32 examples, 0 failures, 1 pending

  real    0m43.649s
  user    0m12.849s
  sys     0m1.260s

Compare that with capybara-webkit:

  Finished in 12.9 seconds
  32 examples, 0 failures, 1 pending

  real    0m19.990s
  user    0m16.098s
  sys     0m1.625s

Note that to install the Gem you need the Qt libraries. Don't go down
the recommended route of installing them via Homebrew; Homebrew is evil,
as far as I can tell, wanting to do crazy shit with the permissions on
the filesystem just to avoid using sudo for installs. (You're doing it
wrong; if you want to avoid using sudo, you install somewhere that
doesn't require elevated privileges, like your home directory, rather
than messing with the permissions on a shared directory.)

There's a binary installer for Mac OS X anyway; just download it and
install from here:

  http://qt.nokia.com/downloads/downloads#qt-lib

Be sure to grab just the Qt libraries; you don't need the full SDK.

Signed-off-by: Wincent Colaiuta <win@wincent.com>