Modernize attribute whitelist handling
So, Rails 4 is probably going to change all this for us:

  https://gist.github.com/1974187 (Yehuda Katz's take)
  https://gist.github.com/1975644 (DHH's take)

but in the meantime, I need to do someting about the way POST params are
being handled in light of the new:

  config.active_record.whitelist_attributes = true

setting.

Use the almost-certain-to-be-deprecated-or-removed `:as` options to
provide role-based whitelist specification. Despite that the fact that
it's on the way out, it has resulted in objectively better/cleaner code.

Signed-off-by: Wincent Colaiuta <win@wincent.com>