script/deploy: big refactor using `run` function
Taking advantage of the fact that we have `ControlMaster` and
`ControlPersist` settings in our `~/.ssh/config`, we can get rid of the
very long strings of "&&"-ed remote commands and instead fire off
separate `ssh` commands for each task, with no real overhead.

This commit adds a `run` helper function to abstract this pattern away,
and give us some niceties for "free", such as a log/deploy.log file.

It is almost certainly not what a Bash veteran would have written, but
it works, and the script is arguably more readable now.

The next big fish to fry here will be involving a local mirror repo for
the initial clone, which should speed up things considerably.

Signed-off-by: Wincent Colaiuta <win@wincent.com>