Fix float issues on box-style comment boxes (issues, forums)
Not sure why I hadn't noticed this before, but here's the fix anyway.

I'm not very happy with where the CSS is at right now. As much as I
preferred the old Sass syntax (because it was what I was used to), now
that I have used the SCSS syntax at work I am warming to the latter. I'm
also painfully aware that I am abusing mixins where classes should
actually be used.

Fixes:

  https://wincent.com/issues/1992

Signed-off-by: Wincent Colaiuta <win@wincent.com>