Make sure separator extends to bottom in box-style comments
I don't like the use of the negative margin hack here, but it seems the
most straightforward way to get the border to always extend to the
bottom.

Basically, we draw the border on the right edge of the `cite` element,
and on the left edge of the `blockquote` element. The negative margin
makes the two borders overlap, which in turn means that the border will
extend to the bottom regardless of which element is taller.

Along the way dropped some unnecessary `px` units from zero margins.

Signed-off-by: Wincent Colaiuta <win@wincent.com>