Add style guide
At this moment it is so minimal and primitive as to almost be a joke,
but I wanted to have at least a starting point to build on as I clean up
the CSS of the site.

Note that I looked at KSS, but it appears to be overkill for my needs,
and I like to keep the Gemfile as lean as possible so that application
startup remains fast.

Signed-off-by: Wincent Colaiuta <win@wincent.com>