Switch to new CloudFront hosts
The first round weren't any good because Amazon was requesting the
assets via HTTP, which nginx then dutifully redirected to HTTPS, and
CloudFront cached the redirect response, not the asset on the other side
of the redirect.

Invalidating objects is a royal pain, so let's just create new
distributions, deploy this code, and then retire the old ones.

Signed-off-by: Wincent Colaiuta <win@wincent.com>