Bump base font size
The text is a little small all round, although it looks nice. Let's just
bump it up by 5%. We could go further, but that will require deeper
adjustments (ie. reducing the font size in certain places, like the
tweets sidebar on the blog, and the tables in the issue tracker).

Signed-off-by: Wincent Colaiuta <win@wincent.com>