Spritify pagination icons
As an initial experiment with Compass spritification, let's tackle a
small section of the application: pagination.

This required a bit of restructuring as sprites are by definition
background images and for the paginator we were previously using inline
images.

I ended up going with a ul for the links with nested spans. This was the
simplest option. I dabbled with pseudo before and after selectors with
content elements, but that would have been quite messy with respect to
clickability (for the psuedo element to be clickable, it would itself
have to either contain a link, or the whole thing would need to be
nested inside a link).

Signed-off-by: Wincent Colaiuta <win@wincent.com>