Move Google Web Fonts into separate style link tag
Avoiding @import as suggested by Page Speed.

Note that I put the style URL in the APP_CONFIG Yaml file so that it
could be "easily" gotten at even from the script/static script, thus
keeping things DRY (even if it is a little odd to store this kind of
thing in a Yaml file).

Signed-off-by: Wincent Colaiuta <win@wincent.com>