Move JS to the bottom of the page
This is not as ideal as truly deferring the parsing of JavaScript,
because the browser won't fire the onload event until the JS is all
downloaded and parsed; this in turn means that the loading indicator
won't go away until the very end.

Nevertheless, it is still a minor win because it lets the browser do its
initial layout and rendering early on, without pausing to download and
parse JS along the way.

I moved the `content_for :javascript` sections down to the bottom of
each template as a visual reminder about the execution order.

Note the reason why we can't just drop in a `defer="defer"` attribute on
the script tags is that we have an ordering dependency:

  - application.js must load first, including jQuery
  - our dynamic JS usually depends on jQuery, so can must load after
    that; for this `defer="defer"` would be fine
  - the catch comes in with our inline JS; early IE and Firefox used to
    allow you to specify `defer="defer"` on inline script elements too,
    but that changed in the HTML5 spec; this in turn means you can't
    depend on jQuery in your inline JS, which is a deal-breaker

So, a deeper look at things will be required. I am thinking something
along the lines of Facebook's Primer library (a handful of lines of JS
in the head element to handle events while waiting for the rest of the
JS to load) or perhaps more use of data attributes to signal to the JS
what it should be doing when it finally loads.

Signed-off-by: Wincent Colaiuta <win@wincent.com>