Add google_web_fonts config to sample app_config.yml
I just had a few minutes of downtime because I forgot that
app_config.yml isn't actually version-controlled, which meant that
production didn't have the necessary "google_web_fonts" setting.

In hindsight, this was evidently not the best place for the setting.
I'll move it if a more suitable choice becomes evident.

Signed-off-by: Wincent Colaiuta <win@wincent.com>