Set gzip_http_version to 1.0 for the benefit of CloudFront
According to the docs:

  http://docs.amazonwebservices.com/AmazonCloudFront/latest/DeveloperGuide/ServingCompressedFiles.html

CloudFront makes requests using HTTP/1.0, but nginx's gzip_http_version
setting defaults to 1.1, meaning that stuff doesn't get compressed.

There are some gotchas in the nginx docs:

  http://wiki.nginx.org/HttpGzipModule#gzip_http_version

warning of the perils of using this setting; namely:

  When HTTP version 1.0 is used, the Vary: Accept-Encoding header is not
  set. As this can lead to proxy cache corruption, consider adding it
  with add_header. Also note that the Content-Length header is not set
  when using either version. Keepalives will therefore be impossible
  with version 1.0, while for 1.1 it is handled by chunked transfers.

Regrettably, it seems like whether the content that CloudFront ends up
distributing is gzipped ultimately depends on whether the first user
agent that happens to requests it includes the appropriate
'Accept-Header'. I fear that if a user without that header set makes the
first request, the uncompressed asset gets cached forever, which is most
unfortunate and basically nullifies the benefit of using the CDN in the
first place (because compression can reduce files massively, to about a
third of their original size). Also, it increases the cost of the CDN,
because its pricing is volume-based.

Signed-off-by: Wincent Colaiuta <win@wincent.com>