Compress bg.jpg more aggressively
It's the single largest asset on most pages, so we gain a lot by
slashing its size.

This was saved at "medium" (30) quality in Photoshop, reducing it from
around 200k to around 40k. Visual quality is evidently degraded, but as
this is a background element, I expect most people won't notice.

Signed-off-by: Wincent Colaiuta <win@wincent.com>