Add new version of libv8 gem to vendor/cache
It's possible that this new Gem appeared in my vendor/cache this evening
as tonight was the first time I've run `bundle` since updating to
Mountain Lion; but that seems a little unlikely as I updated around 25
July 2012 (release day), and I seem to recall `bundle` updating since
then.

It might be because I updated my Ruby version locally tonight.

Signed-off-by: Wincent Colaiuta <win@wincent.com>