Simplify setUpLoginLogoutLinks
Not that I consider this to be a security concern, but this really
belongs in the session and not some random cookie. Of course, the
downside is that having it in the session means I can't get at it from
JS running on static pages so for now I will leave it as-is.

Signed-off-by: Wincent Colaiuta <win@wincent.com>