Add Object.subclass
This will be a dependency of other JS that I will be rolling out in the
future, so it goes near the top of the manifest. Fortunately, we can
rely on Sprockets to do the right thing and only include the "class.js"
file once, even though it is subsumed in the `require_tree .` directive
in the application manifest.

Changes from the version in "Secrets of a JavaScript Ninja":

  - add some additional local variables with names to make some of the
    magic more obvious
  - when possible, blow up if we detect a call to `this._super()` when
    no such method exists (this will only be possible in browsers which
    support decompilation)
  - incur a speed hit by using `try`/`finally` so as to ensure that
    `this._super` is always restored to its prior value

There is some overhead here (basically, an additional wrapping function
around any override of a superclass function, along with the overhead of
the closure). I am not sure if it will be worth it, especially if we
ever want to roll this out in a performance-critical section.

Nevertheless, it is a nice way to get code-reuse, and a nice way to beef
up my understanding of prototypal inheritance.

Signed-off-by: Wincent Colaiuta <win@wincent.com>