Set body font-size using a calculation
Previously we had a hard-coded value in here of 80%.

A common pattern is to set the body size to 10px, so that 1em is 10px.
This makes doing subsequent calculations easy; for example, if you want
a 13px font you just ask for 1.3em.

In practice this is slightly complicated because inheritance can make
things tricky. Nevertheless, it's a good idea for that sets us up well
for flexible/responsive layout.

In my case, however, I'm using Sass and I'm not working with a designer
who wants to tell me sizes in pixels, so I can just set the body size to
13px (the size I actually want) and forget about easy multiplication.
When I need to resize things, Sass will do the math for me.

Furthermore, I am planning on using modular scales soon, so I won't be
thinking in pixels necessarily anyway.

Signed-off-by: Wincent Colaiuta <win@wincent.com>