Add twitter_id, twitter_id_str to tweets table
Twitter will return both an ID and also an ID string, as noted in the
code comment in the migration.

Not sure if it is useful to keep this around, but may as well as space
is not a concern here and it could come in handy for debugging. Also, it
serves as a flag of stuff that has been pushed to the Twitter API.

Signed-off-by: Wincent Colaiuta <win@wincent.com>