Make JS class inheritance simpler and more explicit
Reading the Internets made me think twice about the class inheritance
pattern I was using.

In order to be robust, it has to catch possible errors and that will
make it go much slower. While I don't expect it to be used in a
performance critical section, it still feels a little odd to knowingly
write something so slow, especially when the slowness is required due to
some extremely magical anonymous function generation and closure
shenanigans and property switching.

Another hint that something was wrong was just how hard it was to make
it all readable; I spent considerable time trying to come up with
temporary variables that would reveal what was going on.

Instead, go for something much simpler, based on an idea here:

  https://github.com/rauschma/class-js

The simplicity of the code speaks for itself, and no verbose variable
names or runtime introspection are required. Also, we're not adding any
methods to Object (although we are created a new property on the global
window context, window.Class).

Differences from the linked implementation:

  - use init() instead of constructor()
  - make init() optional, providing a simple default implementation that
    calls through to superclass (if present)

Signed-off-by: Wincent Colaiuta <win@wincent.com>