Add Rake task to build custom jQuery 2.0.1 build
Update to 2.0.1, excluding unnecessary deprecated functionality and
event aliases.

Requires some prerequisites as noted in the Rake task. As I don't want
to rely on these in the production environment, I'm committing the built
product to the repo.

(In hindsight, this means creating a mirror of jQuery was unnecessary,
but I'll leave it as-is for now in case I ever want to hack on it.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>