Add first cut mobile nav
This is a "first cut" because, while it works, it is incomplete. For
example, I currently have it hooked up to activate on clicking the
"Wincent" name in the nav bar; I'll need to replace this with a proper
icon. It also needs to be made conditional on media queries.

This is the first use of the "Class", er, class. I'm ready to start
using Backbone soon, but didn't want to roll out something too heavy
when 50 lines (unminified) of JavaScript would do.

Signed-off-by: Wincent Colaiuta <win@wincent.com>