Fix spec failures and deprecation warnings caused by Rails 4
Fixed up a bunch of failures, and did some clean-up along the way (ie.
converting hashes to Ruby 1.9 syntax). Causes of failures and warnings
included:

  - update forms now use PATCH instead of PUT method
  - addition of mobile nav made Capybara search form selector ambiguous
  - sweepers behave differently outside of controllers now, making
    `#expire_fragment` useless and explosive
  - lots of ActiveRecord deprecations
  - `#truncate` now always returns an HTML-safe string, escaping
    whatever is passed in, unless overridden (we now override)
  - many other things, which I would write about, but am too tired to do
    so after years of chasing Rails

Of all these things, the caching changes are the most vexing; I think
the moral of the story is to not use sweepers at all; DHH wants you to
use "Russian Doll" caching in memcache.

Signed-off-by: Wincent Colaiuta <win@wincent.com>