Remove Atom feeds and page caching
Desperate times call for desperate measures.

With my other projects and my day job taking up all my time, this cruft
is getting in the way of me shipping the Rails 4 upgrade. I really just
want to get this to a nice place and then switch the whole thing over to
Go with a rich client-side front-end... I think.

In any case, I have to reduce the surface area here. The whole point of
rolling my own app was for it to be maintainable, but with every Rails
release breaking/changing core features (caching, sweeping, strong
params etc), I can't afford to be chasing this stuff any more.

Atom and RSS are dead enough that I don't think it matters.

Signed-off-by: Wincent Colaiuta <win@wincent.com>