Switch comment bubbles and flash icons to font-based vectors
Now that we aren't caching any more, we can dispense with the fancy
"cache-friendly flash", which had a security hole in it anyway (eval'ing
use-supplied input; in this case, cookie contents).

As a bonus, switched off the bold (hat tip: Jony Ive).

Also replaced the comment bubble image with a font based one.

Signed-off-by: Wincent Colaiuta <win@wincent.com>