Solve asset compilation problems "once and for all"
Let's just do asset compilation locally. If Rails can't be trusted,
better to keep a close eye on it. (Keep your friends close and your
enemies closer.)

As I am feeling bold and reckless, let's not only do it locally, but
switch to Google's Closure Compiler (yay!) which I consider the best JS
minifier out there, with a lot of other juicy benefits that we can look
at getting into later.

This is sweet, as it allows us to get rid of a lot of cruft from the
Gemfile, including the ever-annoying libv8 which requires an absolute
song-and-dance any time I "upgrade".

With this set-up asset-compilation is ridiculously fast. Takes about 10
seconds.

One thing to note: the Rails docs are out of date; the
`initialize_on_precompile` setting is gone and is now the default. I
did, however, need to put a dummy production config in place in my local
database.yml to get it to work.

Dev assets are now "served" out of /dev-assets and production assets
continue to come from /assets. Note that this regrettably required me to
monkey-patch the Bower-managed copy of Font Awesome. I might look at
tackling that in a less invasive manner in a separate commit, but either
way, it is going to suck a bit.

Still to do: make the deploy script handle compilation automatically;
for now this has to be done manually any time I update an asset.

Signed-off-by: Wincent Colaiuta <win@wincent.com>