Deal with .gitignore shenanigans
`git write-tree` won't work with ignored files and there is no
`--no-ignore-untracked` option or similar.

Likewise, `git add-index` would allow us to explicitly list the files we
want to add, but it won't work with a directory param.

So, here we hack around things by removing the .gitignore file and then
doing a `git add` call; it's porcelain, not plumbing, but so be it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>