Fix release note generation
Release note generation was busted because we previously assumed only
commits were ever tagged; now that we're also tagging asset trees that
was no longer the case.

The first hunk in this diff isn't actually required, but it makes
explicit that we tag all deploys with a tag of the format "0.*". (It's
not required because `git rev-list` only lists commits.)

The second hunk is required. It makes sure that we don't confuse asset
tags (tags of trees) with deploy tags (tags of commits).

Signed-off-by: Wincent Colaiuta <win@wincent.com>