Make deploys faster by using a local mirror
Instead of cloning the entire repo every time, we maintain a local bare
mirror that we update, and then do a very-fast local clone (about 4-6
seconds vs 1m 15s for a network clone).

We create the mirror on the first deploy if it doesn't exist already, so
the first deploy will be as slow as before but subsequent ones will be
faster.

For safety (or paranoia), we remove the "origin" remote from the deploy
clones to prevent accidental push attempts. (Note "origin" here is the
local mirror, so we're really just trying to keep that repo in a
pristine state and avoid corruption.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>