Fix stale manifest usage
If there are multiple asset manifest files in a directory, Sprockets
will use the first it finds. If there are none, it will generate a new
one each time:

  https://github.com/sstephenson/sprockets/blob/eb84c414b76850af9b51bc495a52fb15d6ad24e3/lib/sprockets/manifest.rb#L46

What this means for us in practice is:

  1. Each time we generate assets we get a new randomly-generated digest
     (because each time we compile we do so into a brand new, empty
     directory).
  2. Each time we deploy, our app will use the "first" manifest it
     finds, rather than the latest one we deployed.

This "fix" applied here is to seed the precompilation directory with an
empty JSON file so that it always uses the same digest. The other "fix"
would be to blow away old manifests each time. Neither is particularly
awesome, but I went with the former as it seems less destructive.

Signed-off-by: Wincent Colaiuta <win@wincent.com>