Speed up scrolling on mobile
The application menu was causing a swathe of the screen to be repainted
during scrolling, even when the menu was closed.

My first attempt at fixing this had me moving the width onto the
.menu-open class. This fixed the performance, but added a glitch on
opening where the width of the app content would shrink before the
transition began, leading to a glitchy appearance.

By letting the width get calculated by the contents of the div (the
.menu-inner div), we get good performance and no glitches. The final
piece of the puzzle is to use a margin instead of an explicit width to
make room for hiding the webkit scrollbar.

Signed-off-by: Wincent Colaiuta <win@wincent.com>