Disable user scaling
This is an attempt to make sure that the browser doesn't zoom on
focussing input fields in the global nav on mobile. As a bonus, it will
apparently make the site more responsive on Android by removing the
300ms delay that is imposed after tapping (the OS waits to see if any
given tap is a double tap which could be used to zoom).

Signed-off-by: Wincent Colaiuta <win@wincent.com>