Replace editInPlace with declarative style in 2 places
The amount of code is about the same as before (arguably a little
cleaner), but the behavior is now a little different:

  - on the plus side, the hover effect is nicer
  - on the down side, we're not quite at parity because when we edit
    pending tags we show text rather than links

I'll loop back on the latter.

For now, this is good as it allows us to get rid of issues#show.js,
which Rails has been complaining about since the upgrade to the 4.1.0
beta (and the response to Homakov's security disclosure).

Signed-off-by: Wincent Colaiuta <win@wincent.com>