Fix edit_link_path on links#index
I was tempted to remove this button seeing as you can edit in-line just
fine, but at the same time, it feels reasonable to keep a links#edit
view around and we may as well link to it.

So, switching to id-based lookup for the same reason that I did the same
in commit 88ff83a with the Ajax "contenteditable" elements.

Signed-off-by: Wincent Colaiuta <win@wincent.com>