Add static stylesheet to react-page app
This is a bit kludgey and I am not sure whether I'll end up coming up
with a better way to do this (for example, I'd like to be able to use
Sass), but it works for now and is less cumbersome than dealing with the
React style stuff.

It's not all perfect though; the / in the text/css attribute value is
being entified, which looks like overzealous XSS prevention to me:

  <link rel="stylesheet" type="text&#x2f;css" href="stuff.css" ...

I am not sure how to get around that.

Signed-off-by: Wincent Colaiuta <win@wincent.com>