Work around Ruby 2.1.0 GC bug
This is kind of alarming, and something which makes me wary of putting
2.1.0 out in production yet, but these specs were failing massively
(hundreds of failures) with messages like:

  Failure/Error: Dir.chdir File.join(File.dirname(__FILE__), '..', '..') do
  Errno::EMFILE:
  Too many open files - getcwd
  # ./spec/meta/whitespace_spec.rb:30:in `chdir'

My first instinct was to swap out the File.readlines call. This alone,
wasn't enough to fix things. By using Rails.root instead of the
File.join call, the problem goes away. (Another way to make it go away
is to include a specific `GC.start` before each example.) Note that
switching to Rails.root alone isn't enough; you need the readlines
change as well.

With that much, the file in isolation passes, but when running the suite
as a whole, we boom again. So I'm adding the GC.start call in as well
for good measure.

In short, this is all scary and smelly, but at least with this change
the suite is basically green again. My best hope is that it is
Spring-related, and so not a threat to production.

Signed-off-by: Wincent Colaiuta <win@wincent.com>