React: only allow valid chars in the tag input
This is a work-in-progress that I won't complete, but wanted it in the
Git log for demo purposes.

As you can see, it turns out to be a lot of code to regulate what goes
into the input field, and I haven't even handled the cursor keys yet, or
backspace or delete, or any of a number of other edge-casey things that
I might not even be aware of yet.

Instead I am going to do a transform at pill-ification time, and use
`text-transform: lowercase` to make things look nice during input.

Signed-off-by: Wincent Colaiuta <win@wincent.com>