Implement better "delete to pop" behavior
This will work whether the input field is empty or populated. Note we
check for the cursor being at the start of the field, and only the
start; in this way, if the cursor is "at the start" but a range is
selected, hitting delete will delete the range rather than the preceding
tag.

Signed-off-by: Wincent Colaiuta <win@wincent.com>