Get autocomplete filtering working
This still isn't quite as clean as I would like, but it is better than
the way it was going to be. To make things work better, I moved the
handling of state up into the parent. This feels like "The React Way"
(to let state flow from parent to child).

I might look at using the "classSet" add-on to clean up some of the
manual class management[0].

[0]: http://facebook.github.io/react/docs/class-name-manipulation.html

Signed-off-by: Wincent Colaiuta <win@wincent.com>