Add sample tags.json file
This was generated with:

  Tag.where('taggings_count > 0')
     .order('taggings_count DESC')
     .to_json(only: [:name, :taggings_count], root: false)

in my development environment (which was seeded with some production
data, but has likely been mutated since then and will have some junk in
it).

server.js will happily serve this us, so we can use it as a starting
point for adding proper Ajax autocomplete.

Signed-off-by: Wincent Colaiuta <win@wincent.com>