Pull autocomplete recommendations down via Ajax
No error handling at this point, and I am not sure any is required, as
we can just degrade gracefully when we can't get the completions (eg.
just don't show autocomplete).

Note that I am not using jQuery in this project because I've found that,
so far, React's "declarative style, data-driven approach" means that I
haven't had to do any complex event handling or DOM manipulation. It
might still be nice to have jQuery around (or a custom build of it) for
just the Ajax parts.

The JSON payload here has taggings count in it, so we might want to
display that in the UI at some point (or maybe not; it might just be
clutter). If we decide not to display it, we can slim down the payload
and make things faster.

Signed-off-by: Wincent Colaiuta <win@wincent.com>