Reduce size of JSON payload
The old version was generated with:

  Tag.where('taggings_count > 0')
      .order('taggings_count DESC')
      .to_json(only: [:name, :taggings_count], root: false)

This time we just do:

  Tag.where('taggings_count > 0')
     .order('taggings_count DESC')
     .pluck(:name)
     .to_json

This cuts the payload down to about 25% of its original size, which will
be a win for responsiveness, and to be honest, I don't think we'll
benefit from cluttering up the UI with tag count info. (But will need to
make sure we have a sane query plan on that query; the table is
basically index-free at this point).

Signed-off-by: Wincent Colaiuta <win@wincent.com>