Add an index on `tags` for faster autocomplete
With the old schema, our explain output for:

  EXPLAIN SELECT name
          FROM tags
          WHERE taggings_count > 1
          ORDER BY taggings_count DESC \G

looks like this:

             id: 1
    select_type: SIMPLE
          table: tags
           type: ALL
  possible_keys: NULL
            key: NULL
        key_len: NULL
            ref: NULL
           rows: 416
          Extra: Using where; Using filesort
  1 row in set (0.01 sec)

With the new index, we get this:

             id: 1
    select_type: SIMPLE
          table: tags
           type: index
  possible_keys: index_tags_on_taggings_count_and_name
            key: index_tags_on_taggings_count_and_name
        key_len: 773
            ref: NULL
           rows: 539
          Extra: Using where; Using index
  1 row in set (0.00 sec)

This should ensure our tag autocomplete stays as fast as possible.

One doubt I have is the change to the schema.rb file; I fear this might
be a MariaDB-specific change, which will be incompatible with the MySQL
version in production. Would be nice to update production sooner rather
than later, although I don't want to get bogged down in Ansible stuff
right now...

Signed-off-by: Wincent Colaiuta <win@wincent.com>