Scroll autocomplete if required during keyboard navigation
Make sure the selected element is visible on the screen when
navigating up and down using keyboard input.

Note that the hover-to-select effect here means that you can position
the mouse just right during scrolling and have it scroll automatically
(because the selected element is not fully visible, but when we scroll
to make it visible, the mouse is now hovering over a new element, which
selects it and triggers another scroll...)

I played with some different values of "slop" to make this problem go
away, but it made the keyboard navigation feel janky, so I think it's
best to leave it as-is. As it stands, it is not too obnoxious, and very
edge casey so most people won't run into it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>