Sketch out handling of saved vs unsaved resources
For a saved issue, we can create and destroy tagging instances on the
fly if we want. One way we can do that is with a resourceURL prop, to
which we'll submit POST (create) events and DELETE (destroy) events.
(The sample URL shown here assumes a nested "tags" resource, although in
reality it is a faux-tags resource, partly a "taggings" resource.)

For an unsaved issue, we can't create or destroy tagging instances on
the fly. In that case, we want to include a hidden input field with the
pending tag names. Note that if we don't specify a resourceName (like we
do when we instead pass a resourceURL) React won't insert an empty
"name" attribute into the DOM. (Truth is, it probably should even insert
the input element in that case.)

I am not sure if this is the right way to go (it is decidely non-viewy)
but I'll let it sit and think on it some more.

Signed-off-by: Wincent Colaiuta <win@wincent.com>